requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyMySQL==1.1.2
pytest==8.4.2
python-dotenv==1.1.1
requests==2.32.5
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0
